requirements.txt
abnf==1.1.1
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.8.1
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
beautifulsoup4==4.11.1
bidict==0.22.0
billiard==3.6.4.0
bitarray==2.6.0
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==38.0.1
cytoolz==0.12.0
Deprecated==1.2.13
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
Flask==2.2.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SocketIO==5.2.0
Flask-SQLAlchemy==2.5.1
frozenlist==1.3.1
greenlet==1.1.2
gunicorn==20.1.0
hexbytes==0.3.0
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
itsdangerous==2.0.1
Jinja2==3.1.2
jsonschema==4.16.0
kombu==5.2.4
lru-dict==1.1.8
Mako==1.2.2
MarkupSafe==2.1.1
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
prompt-toolkit==3.0.30
protobuf==3.20.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pycryptodome==3.15.0
PyJWT==2.5.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
pytest-celery==0.0.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-engineio==4.3.4
python-socketio==5.7.1
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
rfc3987==1.3.8
rlp==2.0.1
siwe==2.0.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.40
tomli==2.0.1
toolz==0.12.0
urllib3==1.26.11
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.30.0
websockets==9.1
Werkzeug==2.2.2
wrapt==1.14.1
yarl==1.8.1
zipp==3.8.1
